safari-600.7-branch/LayoutTests/svg/animations/svgboolean-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test 'to' animation of SVGBoolean.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is true
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS successfullyParsed is true

TEST COMPLETE
